Add an optimized text rendering example using a texture atlas.
This example renders exactly the same text as the first text example.  However,
it creates a texture atlas of all the characters, so that a texture upload
is done only once for each font/size combination, and so that all
characters in a line of text can be drawn with a single glDrawArrays() command.