Add a version of glescraft that uses a geometry shader and 3D textures.
A face of a cube can be represented by two vertices. This allows the geometry
shader to generate a quad and calculate the surface normal. In the VBO, the w
component of every second vertex can be used for something different than the
texture coordinate. It is now used to contain lighting information.

Instead of a 2D texture atlas, this version also uses a 3D texture, which
removes the need to implement texture wrapping in the fragment shader.