Add a simple ray casting method to the object selection tutorial.
We need to find two points on the ray determined by the mouse cursor position.
We know that every camera ray goes through the center of the camera, so that is
one point. We can get another point by unprojecting the mouse cursor
coordinates with any depth value. GLM's closestPointOnLine() function to find
the point on the ray that is closest to the center of a box, and then calculate
the distance between that point and the box's center to find out how close it
is to the ray. If it is less than the "diameter" of the box, we assume the ray
passes through the box. If more than one box lies on the ray, we select the
one closest to the camera.