Add a function to create a program with a geometry shader.
It is slightly more complicated than just compiling a bunch of shaders and
linking them; some parameters for the geometry shader must be set before the
program is linked. Note that this is for GLSL 1.20 with the geometry shader
extension; GLSL 1.50, which support geometry shaders by default, requires you
to specify the input and output format in the shader source code itself.